+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Change Log
 
+## 0.0.4
+
+- `options.fil`
+
+- Some breaking changes to arguments
+  - `options.path`
+  - `options.native`
+  - `options.canChangeFolder`
+  - `output.default`
+  - `output.defaultPath`
+
 ## 0.0.3
 
 - Some breaking changes to arguments